Some refactor of project
